chore: resolve merge conflicts with v2/develop
