+++ frontend/src/features/admin-form/create/builder-and-design/BuilderAndDesignContent/StartPageView.tsx
@@ -34,6 +34,7 @@ export const StartPageView = () => {
         estTimeString={estTimeString}
         titleBg={titleBg}
         titleColor={titleColor}
+        showHeader={true}
         loggedInId={
           form?.authType !== FormAuthType.NIL ? 'S8899000D' : undefined
         }
